Prepare inline text changes for review
